[merge] Move layout code into merge.layout
